background subtraction and convoluted exponential with fixed parameters implemented in the maximum likelihood fit
Done:
-separates decay times in signal region/background sideband according
to DeltaM.
-calculates weight
-takes into account the convolution with a gaussian (at fixed
parameters) and applies weight factor when applicable in the maximum
likelihood.
-numerically computes the first and second derivatives of the negative
log likelihood with respect to tau
-finds the root of the derivative of the negative log likelihood with
Newton raphson method
-computes a statistical uncertainty on the root calculated above

Why it might crash:
-we need to obtain the negative log likelihood (after taking the
product with respect to each t_i and the log) as a function of tau. If
tau (an array) is substituted before, the product might not give the
result expected.
-consequence of said above, the differentiation with respect to tau
might not work.
-the neg log likelihood function is built awkwardly to try to tackle
the problem of tau being an array and the compatibility of the product
with respect to t_i.

To do:
-check that everything is working and gives expected results.
-special care to the function Neg log likelihood
-check the differentiation (and that it’s differentiating with the
right variable)
-print results (weight, lifetime, uncertainty) in special file
-implement the lifetime in the fit
-plot the decay distribution with the weight factor

Lifetime.py
-l. 318 in the definition of the convoluted exponential, make sure to
add the factor of l/2 in the exponential exponent
-end of file: change the weight calculation to wb = - (na/nb) and make
sure na is right